tried to avoid int overflow ?

+++ lib/stb_image.h
@@ -1778,7 +1778,7 @@ static stbi__uint16 *stbi__convert_format16(stbi__uint16 *data, int img_n,
   if (req_comp == img_n) return data;
   STBI_ASSERT(req_comp >= 1 && req_comp <= 4);
 
-  good = (stbi__uint16 *)stbi__malloc(req_comp * x * y * 2);
+  good = (stbi__uint16 *)stbi__malloc((size_t)req_comp * x * y * 2);
   if (good == NULL) {
     STBI_FREE(data);
     return (stbi__uint16 *)stbi__errpuc("outofmem", "Out of memory");
@@ -6365,7 +6365,7 @@ static void *stbi__psd_load(stbi__context *s, int *x, int *y, int *comp,
     out = (stbi_uc *)stbi__malloc_mad3(8, w, h, 0);
     ri->bits_per_channel = 16;
   } else
-    out = (stbi_uc *)stbi__malloc(4 * w * h);
+    out = (stbi_uc *)stbi__malloc((size_t)4 * w * h);
 
   if (!out) return stbi__errpuc("outofmem", "Out of memory");
   pixelCount = w * h;
@@ -6680,7 +6680,7 @@ static void *stbi__pic_load(stbi__context *s, int *px, int *py, int *comp,
 
   // intermediate buffer is RGBA
   result = (stbi_uc *)stbi__malloc_mad3(x, y, 4, 0);
-  memset(result, 0xff, x * y * 4);
+  memset(result, 0xff, (size_t)x * y * 4);
 
   if (!stbi__pic_load_core(s, x, y, comp, result)) {
     STBI_FREE(result);
@@ -6988,12 +6988,12 @@ static stbi_uc *stbi__gif_load_next(stbi__context *s, stbi__gif *g, int *comp,
     }
 
     // background is what out is after the undoing of the previou frame;
-    memcpy(g->background, g->out, 4 * g->w * g->h);
+    memcpy(g->background, g->out, (size_t)4 * g->w * g->h);
   }
 
   // clear my history;
   memset(g->history, 0x00,
-         g->w * g->h); // pixels that were affected previous frame
+         (size_t)g->w * g->h); // pixels that were affected previous frame
 
   for (;;) {
     int tag = stbi__get8(s);
@@ -7134,7 +7134,7 @@ static void *stbi__load_gif_main(stbi__context *s, int **delays, int *x, int *y,
         stride = g.w * g.h * 4;
 
         if (out) {
-          void *tmp = (stbi_uc *)STBI_REALLOC(out, layers * stride);
+          void *tmp = (stbi_uc *)STBI_REALLOC(out, (size_t)layers * stride);
           if (NULL == tmp) {
             STBI_FREE(g.out);
             STBI_FREE(g.history);
@@ -7146,7 +7146,7 @@ static void *stbi__load_gif_main(stbi__context *s, int **delays, int *x, int *y,
             *delays = (int *)STBI_REALLOC(*delays, sizeof(int) * layers);
           }
         } else {
-          out = (stbi_uc *)stbi__malloc(layers * stride);
+          out = (stbi_uc *)stbi__malloc((size_t)layers * stride);
           if (delays) {
             *delays = (int *)stbi__malloc(layers * sizeof(int));
           }
